+++ src/main/java/http/ContentType.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 import java.util.Arrays;
