+++ src/main/java/http/RequestParser.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 import java.io.BufferedReader;
 import java.io.IOException;
