+++ src/main/java/http/Status.java
@@ -1,4 +1,4 @@
-package webserver;
+package http;
 
 public enum Status {
