+++ src/test/java/common/util/UtilTest.java
@@ -1,4 +1,4 @@
-package util;
+package common.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
